Moved api call for status updates to card component
