+++ src/sass/components/ATATSearch.scss
@@ -2,6 +2,7 @@
   fieldset {
     border-top-right-radius: 0;
     border-bottom-right-radius: 0;
+    top: -4px;
   }
   .v-text-field__details {
     display: none;
